requirements.txt
torch
torchvision
torchaudio
opencv-python
numpy
matplotlib
pillow
scikit-learn